requirements.txt
fastapi[all]
asyncpg
python-dotenv
cryptography
twikit
